{"types":"C","runs":29}

+++ problem17.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class problem17 extends ConsoleProgram{
 	public void run() {
-		double a = readInt("enter number: ");
+		double a = readDouble("enter number: ");
 		
 		int b = (int) a;
 		println(b);
